requirements.txt
praw==7.7.1
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.0
pytz==2023.3
nltk==3.8.1
textblob==0.17.1
beautifulsoup4==4.12.2
aiohttp==3.9.1
json5==0.9.14
